Add vocalpy.energy.signal with meansquared
